// gatsby/types/graphql.ts
export type Maybe<T> = T | null
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /**
   * A date string, such as 2007-12-03, compliant with the ISO 8601 standard for
   * representation of dates and times using the Gregorian calendar.
   */
  Date: any
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any
}

export enum BannerStyle {
  FullWidth = 'FULL_WIDTH',
  Normal = 'NORMAL',
}

export type BannerStyleQueryOperatorInput = {
  eq?: Maybe<BannerStyle>
  ne?: Maybe<BannerStyle>
  in?: Maybe<Array<Maybe<BannerStyle>>>
  nin?: Maybe<Array<Maybe<BannerStyle>>>
}

export type BlogPost = {
  id: Scalars['ID']
  globalBlogPostId: Scalars['ID']
  title: Scalars['String']
  body: Scalars['String']
  excerpt: Scalars['String']
  banner?: Maybe<File>
  bannerStyle?: Maybe<BannerStyle>
  externalLinks: ExternalLinks
  slug: Scalars['String']
  langKey: LangKey
  date: Scalars['Date']
  tags?: Maybe<Array<Scalars['String']>>
  category: Scalars['String']
  categorySlug: Scalars['String']
  tableOfContents: Scalars['JSON']
  description?: Maybe<Scalars['String']>
  keywords?: Maybe<Array<Scalars['String']>>
}

export type BlogPostExcerptArgs = {
  pruneLength?: Maybe<Scalars['Int']>
}

export type BlogPostDateArgs = {
  formatString?: Maybe<Scalars['String']>
  fromNow?: Maybe<Scalars['Boolean']>
  difference?: Maybe<Scalars['String']>
  locale?: Maybe<Scalars['String']>
}

export type BlogPostTableOfContentsArgs = {
  maxDepth?: Maybe<Scalars['Int']>
}

export type BlogPostConnection = {
  __typename?: 'BlogPostConnection'
  totalCount: Scalars['Int']
  edges: Array<BlogPostEdge>
  nodes: Array<BlogPost>
  pageInfo: PageInfo
  distinct: Array<Scalars['String']>
  group: Array<BlogPostGroupConnection>
}

export type BlogPostConnectionDistinctArgs = {
  field: BlogPostFieldsEnum
}

export type BlogPostConnectionGroupArgs = {
  skip?: Maybe<Scalars['Int']>
  limit?: Maybe<Scalars['Int']>
  field: BlogPostFieldsEnum
}

export type BlogPostEdge = {
  __typename?: 'BlogPostEdge'
  next?: Maybe<BlogPost>
  node: BlogPost
  previous?: Maybe<BlogPost>
}
// ...
export type BlogPostGroupConnection = {
  __typename?: 'BlogPostGroupConnection'
  totalCount: Scalars['Int']
  edges: Array<BlogPostEdge>
  nodes: Array<BlogPost>
  pageInfo: PageInfo
  field: Scalars['String']
  fieldValue?: Maybe<Scalars['String']>
}
// ...
export type Directory = Node & {
  __typename?: 'Directory'
  sourceInstanceName: Scalars['String']
  absolutePath: Scalars['String']
  relativePath: Scalars['String']
  extension: Scalars['String']
  size: Scalars['Int']
  prettySize: Scalars['String']
  modifiedTime: Scalars['Date']
  accessTime: Scalars['Date']
  changeTime: Scalars['Date']
  birthTime: Scalars['Date']
  root: Scalars['String']
  dir: Scalars['String']
  base: Scalars['String']
  ext: Scalars['String']
  name: Scalars['String']
  relativeDirectory: Scalars['String']
  dev: Scalars['Int']
  mode: Scalars['Int']
  nlink: Scalars['Int']
  uid: Scalars['Int']
  gid: Scalars['Int']
  rdev: Scalars['Int']
  ino: Scalars['Float']
  atimeMs: Scalars['Float']
  mtimeMs: Scalars['Float']
  ctimeMs: Scalars['Float']
  atime: Scalars['Date']
  mtime: Scalars['Date']
  ctime: Scalars['Date']
  /** @deprecated Use `birthTime` instead */
  birthtime?: Maybe<Scalars['Date']>
  /** @deprecated Use `birthTime` instead */
  birthtimeMs?: Maybe<Scalars['Float']>
  blksize?: Maybe<Scalars['Int']>
  blocks?: Maybe<Scalars['Int']>
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}
// ...
export type DirectoryConnection = {
  __typename?: 'DirectoryConnection'
  totalCount: Scalars['Int']
  edges: Array<DirectoryEdge>
  nodes: Array<Directory>
  pageInfo: PageInfo
  distinct: Array<Scalars['String']>
  group: Array<DirectoryGroupConnection>
}
// ...
export type ExternalLinks = {
  __typename?: 'ExternalLinks'
  devto?: Maybe<Scalars['String']>
  canonical?: Maybe<Scalars['String']>
  github: Scalars['String']
}

export type ExternalLinksFilterInput = {
  devto?: Maybe<StringQueryOperatorInput>
  canonical?: Maybe<StringQueryOperatorInput>
  github?: Maybe<StringQueryOperatorInput>
}

export type File = Node & {
  __typename?: 'File'
  sourceInstanceName: Scalars['String']
  absolutePath: Scalars['String']
  relativePath: Scalars['String']
  extension: Scalars['String']
  size: Scalars['Int']
  prettySize: Scalars['String']
  modifiedTime: Scalars['Date']
  accessTime: Scalars['Date']
  changeTime: Scalars['Date']
  birthTime: Scalars['Date']
  root: Scalars['String']
  dir: Scalars['String']
  base: Scalars['String']
  ext: Scalars['String']
  name: Scalars['String']
  relativeDirectory: Scalars['String']
  dev: Scalars['Int']
  mode: Scalars['Int']
  nlink: Scalars['Int']
  uid: Scalars['Int']
  gid: Scalars['Int']
  rdev: Scalars['Int']
  ino: Scalars['Float']
  atimeMs: Scalars['Float']
  mtimeMs: Scalars['Float']
  ctimeMs: Scalars['Float']
  atime: Scalars['Date']
  mtime: Scalars['Date']
  ctime: Scalars['Date']
  /** @deprecated Use `birthTime` instead */
  birthtime?: Maybe<Scalars['Date']>
  /** @deprecated Use `birthTime` instead */
  birthtimeMs?: Maybe<Scalars['Float']>
  blksize?: Maybe<Scalars['Int']>
  blocks?: Maybe<Scalars['Int']>
  /** Copy file to static directory and return public url to it */
  publicURL?: Maybe<Scalars['String']>
  childImageSharp?: Maybe<ImageSharp>
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
  childMdx?: Maybe<Mdx>
}
// ...
export type FileConnection = {
  __typename?: 'FileConnection'
  totalCount: Scalars['Int']
  edges: Array<FileEdge>
  nodes: Array<File>
  pageInfo: PageInfo
  distinct: Array<Scalars['String']>
  group: Array<FileGroupConnection>
}
// ...
export type ImageSharp = Node & {
  __typename?: 'ImageSharp'
  fixed?: Maybe<ImageSharpFixed>
  /** @deprecated Resolutions was deprecated in Gatsby v2. It's been renamed to "fixed" https://example.com/write-docs-and-fix-this-example-link */
  resolutions?: Maybe<ImageSharpResolutions>
  fluid?: Maybe<ImageSharpFluid>
  /** @deprecated Sizes was deprecated in Gatsby v2. It's been renamed to "fluid" https://example.com/write-docs-and-fix-this-example-link */
  sizes?: Maybe<ImageSharpSizes>
  original?: Maybe<ImageSharpOriginal>
  resize?: Maybe<ImageSharpResize>
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}
// ...
export type ImageSharpConnection = {
  __typename?: 'ImageSharpConnection'
  totalCount: Scalars['Int']
  edges: Array<ImageSharpEdge>
  nodes: Array<ImageSharp>
  pageInfo: PageInfo
  distinct: Array<Scalars['String']>
  group: Array<ImageSharpGroupConnection>
}
// ...
export enum LangKey {
  En = 'EN',
  PtBr = 'PT_BR',
}

export type LangKeyQueryOperatorInput = {
  eq?: Maybe<LangKey>
  ne?: Maybe<LangKey>
  in?: Maybe<Array<Maybe<LangKey>>>
  nin?: Maybe<Array<Maybe<LangKey>>>
}

export type Mdx = Node & {
  __typename?: 'Mdx'
  rawBody: Scalars['String']
  fileAbsolutePath: Scalars['String']
  frontmatter?: Maybe<MdxFrontmatter>
  body: Scalars['String']
  excerpt: Scalars['String']
  headings?: Maybe<Array<Maybe<MdxHeadingMdx>>>
  html?: Maybe<Scalars['String']>
  mdxAST?: Maybe<Scalars['JSON']>
  tableOfContents?: Maybe<Scalars['JSON']>
  timeToRead?: Maybe<Scalars['Int']>
  wordCount?: Maybe<MdxWordCount>
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
  childMdxBlogPost?: Maybe<MdxBlogPost>
}

export type MdxExcerptArgs = {
  pruneLength?: Maybe<Scalars['Int']>
  truncate?: Maybe<Scalars['Boolean']>
}

export type MdxHeadingsArgs = {
  depth?: Maybe<HeadingsMdx>
}

export type MdxTableOfContentsArgs = {
  maxDepth?: Maybe<Scalars['Int']>
}

export type MdxBlogPost = BlogPost &
  Node & {
    __typename?: 'MdxBlogPost'
    id: Scalars['ID']
    globalBlogPostId: Scalars['ID']
    title: Scalars['String']
    banner?: Maybe<File>
    bannerStyle?: Maybe<BannerStyle>
    externalLinks: ExternalLinks
    slug: Scalars['String']
    langKey: LangKey
    date: Scalars['Date']
    tags?: Maybe<Array<Scalars['String']>>
    category: Scalars['String']
    categorySlug: Scalars['String']
    description?: Maybe<Scalars['String']>
    keywords?: Maybe<Array<Scalars['String']>>
    tableOfContents: Scalars['JSON']
    excerpt: Scalars['String']
    body: Scalars['String']
    parent?: Maybe<Node>
    children: Array<Node>
    internal: Internal
  }

export type MdxBlogPostDateArgs = {
  formatString?: Maybe<Scalars['String']>
  fromNow?: Maybe<Scalars['Boolean']>
  difference?: Maybe<Scalars['String']>
  locale?: Maybe<Scalars['String']>
}

export type MdxBlogPostTableOfContentsArgs = {
  maxDepth?: Maybe<Scalars['Int']>
}

export type MdxBlogPostExcerptArgs = {
  pruneLength?: Maybe<Scalars['Int']>
}

export type MdxBlogPostConnection = {
  __typename?: 'MdxBlogPostConnection'
  totalCount: Scalars['Int']
  edges: Array<MdxBlogPostEdge>
  nodes: Array<MdxBlogPost>
  pageInfo: PageInfo
  distinct: Array<Scalars['String']>
  group: Array<MdxBlogPostGroupConnection>
}
// ...
export type MdxConnection = {
  __typename?: 'MdxConnection'
  totalCount: Scalars['Int']
  edges: Array<MdxEdge>
  nodes: Array<Mdx>
  pageInfo: PageInfo
  distinct: Array<Scalars['String']>
  group: Array<MdxGroupConnection>
}
// ...
export type Node = {
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}

export type NodeFilterInput = {
  id?: Maybe<StringQueryOperatorInput>
  parent?: Maybe<NodeFilterInput>
  children?: Maybe<NodeFilterListInput>
  internal?: Maybe<InternalFilterInput>
}

export type NodeFilterListInput = {
  elemMatch?: Maybe<NodeFilterInput>
}

export type PageInfo = {
  __typename?: 'PageInfo'
  currentPage: Scalars['Int']
  hasPreviousPage: Scalars['Boolean']
  hasNextPage: Scalars['Boolean']
  itemCount: Scalars['Int']
  pageCount: Scalars['Int']
  perPage?: Maybe<Scalars['Int']>
  totalCount: Scalars['Int']
}
// ...
export type Query = {
  __typename?: 'Query'
  blogPost?: Maybe<BlogPost>
  allBlogPost: BlogPostConnection
  file?: Maybe<File>
  allFile: FileConnection
  directory?: Maybe<Directory>
  allDirectory: DirectoryConnection
  sitePage?: Maybe<SitePage>
  allSitePage: SitePageConnection
  site?: Maybe<Site>
  allSite: SiteConnection
  imageSharp?: Maybe<ImageSharp>
  allImageSharp: ImageSharpConnection
  mdx?: Maybe<Mdx>
  allMdx: MdxConnection
  mdxBlogPost?: Maybe<MdxBlogPost>
  allMdxBlogPost: MdxBlogPostConnection
  siteBuildMetadata?: Maybe<SiteBuildMetadata>
  allSiteBuildMetadata: SiteBuildMetadataConnection
  sitePlugin?: Maybe<SitePlugin>
  allSitePlugin: SitePluginConnection
}
// ...
export type Site = Node & {
  __typename?: 'Site'
  buildTime?: Maybe<Scalars['Date']>
  siteMetadata?: Maybe<SiteSiteMetadata>
  port?: Maybe<Scalars['Int']>
  host?: Maybe<Scalars['String']>
  pathPrefix?: Maybe<Scalars['String']>
  polyfill?: Maybe<Scalars['Boolean']>
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}

export type SiteBuildTimeArgs = {
  formatString?: Maybe<Scalars['String']>
  fromNow?: Maybe<Scalars['Boolean']>
  difference?: Maybe<Scalars['String']>
  locale?: Maybe<Scalars['String']>
}

export type SiteBuildMetadata = Node & {
  __typename?: 'SiteBuildMetadata'
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
  buildTime?: Maybe<Scalars['Date']>
}

export type SiteBuildMetadataBuildTimeArgs = {
  formatString?: Maybe<Scalars['String']>
  fromNow?: Maybe<Scalars['Boolean']>
  difference?: Maybe<Scalars['String']>
  locale?: Maybe<Scalars['String']>
}

export type SiteBuildMetadataConnection = {
  __typename?: 'SiteBuildMetadataConnection'
  totalCount: Scalars['Int']
  edges: Array<SiteBuildMetadataEdge>
  nodes: Array<SiteBuildMetadata>
  pageInfo: PageInfo
  distinct: Array<Scalars['String']>
  group: Array<SiteBuildMetadataGroupConnection>
}
// ...
export type SiteConnection = {
  __typename?: 'SiteConnection'
  totalCount: Scalars['Int']
  edges: Array<SiteEdge>
  nodes: Array<Site>
  pageInfo: PageInfo
  distinct: Array<Scalars['String']>
  group: Array<SiteGroupConnection>
}
// ...
export type SitePage = Node & {
  __typename?: 'SitePage'
  path: Scalars['String']
  component: Scalars['String']
  internalComponentName: Scalars['String']
  componentChunkName: Scalars['String']
  matchPath?: Maybe<Scalars['String']>
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
  isCreatedByStatefulCreatePages?: Maybe<Scalars['Boolean']>
  context?: Maybe<SitePageContext>
  pluginCreator?: Maybe<SitePlugin>
  pluginCreatorId?: Maybe<Scalars['String']>
  componentPath?: Maybe<Scalars['String']>
}

export type SitePageConnection = {
  __typename?: 'SitePageConnection'
  totalCount: Scalars['Int']
  edges: Array<SitePageEdge>
  nodes: Array<SitePage>
  pageInfo: PageInfo
  distinct: Array<Scalars['String']>
  group: Array<SitePageGroupConnection>
}
// ...
export type SitePlugin = Node & {
  __typename?: 'SitePlugin'
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
  resolve?: Maybe<Scalars['String']>
  name?: Maybe<Scalars['String']>
  version?: Maybe<Scalars['String']>
  pluginOptions?: Maybe<SitePluginPluginOptions>
  nodeAPIs?: Maybe<Array<Maybe<Scalars['String']>>>
  browserAPIs?: Maybe<Array<Maybe<Scalars['String']>>>
  ssrAPIs?: Maybe<Array<Maybe<Scalars['String']>>>
  pluginFilepath?: Maybe<Scalars['String']>
  packageJson?: Maybe<SitePluginPackageJson>
}

export type SitePluginConnection = {
  __typename?: 'SitePluginConnection'
  totalCount: Scalars['Int']
  edges: Array<SitePluginEdge>
  nodes: Array<SitePlugin>
  pageInfo: PageInfo
  distinct: Array<Scalars['String']>
  group: Array<SitePluginGroupConnection>
}
// ...
export type GatsbyCreatePagesQuery = { __typename?: 'Query' } & {
  allBlogPost: { __typename?: 'BlogPostConnection' } & {
    edges: Array<
      { __typename?: 'BlogPostEdge' } & {
        node: { __typename?: 'MdxBlogPost' } & Pick<
          MdxBlogPost,
          | 'id'
          | 'excerpt'
          | 'title'
          | 'slug'
          | 'langKey'
          | 'category'
          | 'categorySlug'
          | 'globalBlogPostId'
          | 'tags'
          | 'body'
        >
      }
    >
  }
}